Only crawl `metadata.json` for the Athena table.
The schema of the transcript JSON data varies too much for the Glue crawler to handle without sampling all records.  And, there is not a lot of value in having all the raw transcript data available in the Athena database.

Instead, we'll use the Athena database more like a card catalog at a library, storing only the metadata.  The transcript JSON will be stored separately, in parallel with the `metadata.json`, and it will be ignored by the Glue crawler.